reducers is separeted on 2 files
