+++ heroes.json
@@ -7,22 +7,16 @@
       "element": "fire"
     },
     {
-      "id": 2,
-      "name": "Карен",
-      "description": "керить",
-      "element": ""
-    },
-    {
-      "id": 3,
-      "name": "sdnfjsdj",
-      "description": "njfskr",
+      "id": "941547bf-e09a-46af-a1b9-b330f73c4d58",
+      "name": "kdsmfl",
+      "description": "las",
       "element": "water"
     },
     {
-      "id": 4,
-      "name": "Jack",
-      "description": "Jack",
-      "element": "wind"
+      "id": "ac1a49f1-140b-4e7e-99e0-918f292e7192",
+      "name": "21 km ",
+      "description": "23km",
+      "element": "fire"
     }
   ],
   "filters": [
