+++ src/components/heroesListItem/HeroesListItem.js
@@ -31,8 +31,8 @@ const HeroesListItem = ({name, description, element, onDelete}) => {
                 <h3 className="card-title">{name}</h3>
                 <p className="card-text">{description}</p>
             </div>
-            <span className="position-absolute top-0 start-100 translate-middle badge border rounded-pill bg-light">
-                <button type="button" className="btn-close btn-close" aria-label="Close" onClick = {onDelete}></button>
+            <span onClick = {onDelete} className="position-absolute top-0 start-100 translate-middle badge border rounded-pill bg-light">
+                <button type="button" className="btn-close btn-close" aria-label="Close"></button>
             </span>
         </li>
     )
