+++ src/reducers/filters.js
@@ -1,0 +1,34 @@
+const initialState = {    
+    filters: [],
+    filtersLoadingStatus: 'idle',
+    activeFilter: 'all',
+}
+
+const filters = (state = initialState, action) => {
+    switch (action.type) {
+        case 'FILTERS_FETCHING': 
+            return {
+                ...state,
+                filtersLoadingStatus: 'loading'
+            };
+        case 'FILTERS_FETCHED':
+            return {
+                ...state,
+                filters: action.payload,
+                filtersLoadingStatus: 'idle'
+            }
+        case 'FILTERS_FETCHING_ERROR':   
+            return {
+                ...state,
+                heroesLoadingStatus: 'error'
+            }
+        case 'ACTIVE_FILTER_CHANGED':
+            return {
+                ...state,
+                activeFilter: action.payload,
+            }
+        default: return state
+    }
+}
+
+export default filters;
