+++ src/store/index.js
@@ -1,6 +1,9 @@
-import { createStore } from 'redux';
-import reducer from '../reducers';
+import { createStore, combineReducers} from 'redux';
 
-const store = createStore(reducer, window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__());
+import heroes from '../reducers/heroes';
+import filters from '../reducers/filters';
+
+const store = createStore( combineReducers({heroes, filters}), 
+    window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__());
 
 export default store;
